Initial commit
This commit adds the a C pwenc library for use with a truenas
APM module and the truenas_pypwenc cpython extension. The purpose
of this library is to provide identical functionality to the
legacy pwenc plugin in the truenas middleware. The library uses
openssl AES-256-CTR to encrypt and base64 encode a given payload
in one direction or base64 decode and decrypt in another direction
using the pwenc seed value specified.